release: version 1.0.8 (versionCode 10)
- Fix AI parsing crash (Double vs Int)
- Update VisionScreen and VisionViewModel to handle Double analysis values
- Add unit test for decimal parsing in GeminiParser
- Increment version to 1.0.8

+++ app/src/main/java/com/healthio/core/ai/FoodAnalysis.kt
@@ -2,12 +2,12 @@ package com.healthio.core.ai
 
 data class FoodAnalysis(
     val foodName: String,
-    val calories: Int,
-    val protein: Int,
-    val carbs: Int,
-    val fat: Int,
-    val fiber: Int, // Added for Net Carbs
-    val insulinScore: Int, // 0-100 scale (Insulin Index)
-    val healthScore: Int, // 1-10
+    val calories: Double,
+    val protein: Double,
+    val carbs: Double,
+    val fat: Double,
+    val fiber: Double,
+    val insulinScore: Double,
+    val healthScore: Double,
     val feedback: String
 )
